+++ win_utils.h
@@ -26,6 +26,7 @@ void ConstrainByBorders(TwmWindow *twmwin, int *left, int width,
                         int *top, int height);
 void WarpToWindow(TwmWindow *t, bool must_raise);
 void send_clientmessage(Window w, Atom a, Time timestamp);
+XWMHints *gen_synthetic_wmhints(TwmWindow *win);
 
 
 #endif /* _CTWM_WIN_UTILS_H */